Moved capabilities into it's own file
